requirements.txt
PyJWT==2.6.0
aiohttp==3.8.4

# Dev dependencies
black==23.1.0
isort==5.12.0
pytest-aiohttp==1.0.4
pytest-mock==3.10.0
pytest==7.2.1
